# client/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Distill",
    "description": "Distill turns unstructured and semi-structured documents into clean, structured data that can\nbe searched and queried.\n\nThe interesting problems are not extraction from a single document, which modern language\nmodels largely solve, but **unification** (turning many mutually disagreeing documents into\none coherent, queryable schema) and **trust** (making every extracted value traceable to\nwhere it came from, and cheap for a person to correct).\n\nAuthentication is a workspace bearer token, issued once at workspace creation. There are no\naccounts.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/workspaces/{workspace_id}/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload one or more documents",
        "operationId": "upload_api_v1_workspaces__workspace_id__documents_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_workspaces__workspace_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/documents/seed": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Load the sample documents",
        "description": "Load the curated sample set. Requirement FR-05.\n\nReads ``samples/manifest.json`` rather than hard-coding filenames, so dropping files\ninto ``samples/`` is the only step needed to change the demo corpus.",
        "operationId": "seed_api_v1_workspaces__workspace_id__documents_seed_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "One document",
        "operationId": "get_document_api_v1_workspaces__workspace_id__documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document identifier.",
              "title": "Document Id"
            },
            "description": "Document identifier."
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete a document",
        "operationId": "delete_document_api_v1_workspaces__workspace_id__documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document identifier.",
              "title": "Document Id"
            },
            "description": "Document identifier."
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/documents/{document_id}/file": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "The original file, with byte-range support",
        "description": "Stream the original. Range requests are supported because pdf.js relies on them.\n\nWithout ranges, opening page one of a fifty megabyte scan downloads all fifty\nmegabytes first, which is the difference between a viewer that feels instant and one\nthat does not.",
        "operationId": "original_file_api_v1_workspaces__workspace_id__documents__document_id__file_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document identifier.",
              "title": "Document Id"
            },
            "description": "Document identifier."
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces/{workspace_id}/documents/{document_id}/pages/{page_index}/image": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "A rendered page image",
        "operationId": "page_image_api_v1_workspaces__workspace_id__documents__document_id__pages__page_index__image_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Document identifier.",
              "title": "Document Id"
            },
            "description": "Document identifier."
          },
          {
            "name": "page_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Index"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workspace Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_v1_workspaces__workspace_id__documents_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "One or more documents."
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_api_v1_workspaces__workspace_id__documents_post"
      },
...
      "UploadResponse": {
        "properties": {
          "accepted": {
            "items": {
              "$ref": "#/components/schemas/UploadedDocument"
            },
            "type": "array",
            "title": "Accepted"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedUpload"
            },
            "type": "array",
            "title": "Rejected"
          }
        },
        "type": "object",
        "title": "UploadResponse",
        "description": "Partial success is the normal case, so accepted and rejected come back together.\n\nOne unreadable file in a drop of twelve must not fail the other eleven, which is\nguarantee 3 in the backend plan and requirement 3.6."
      },
...
    }
  }
}